requirements.txt
# PDFexport - Dependencias Python
# Framework web
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Tareas programadas
APScheduler==3.10.4

# Manipulacion de PDF
PyMuPDF==1.23.7

# Conversion a imagenes
pdf2image==1.16.3
Pillow==10.1.0

# Extraccion de texto
pdfminer.six==20221105

# Generacion de DOCX
python-docx==1.1.0

# HTML a PDF
# WeasyPrint 60.1 requiere pydyf<0.11.0 (en 0.11.0 cambiaron PDF.__init__)
WeasyPrint==60.1
pydyf>=0.5.0,<0.11.0
requests==2.31.0

# Extraccion de tablas de PDF a CSV (Etapa 15)
pdfplumber==0.11.0

# Web scraping (Etapa 16)
beautifulsoup4==4.12.3
trafilatura==1.12.0
markdownify==0.13.1
lxml>=5.2.2

# Utilidades
python-dateutil==2.8.2
python-dotenv==1.0.0
